refactor(dom): CSS data-driven — tabela única de propriedades (css_props!) gera cascade/herança/anim
O ComputedStyle deixou de ser mantido campo-a-campo em 5+ lugares espalhados
(struct + ~100 linhas de merge_over + lista de herança no dom.rs + lista fixa
de styles_differ_animated + interpolate campo-a-campo no anim.rs — listas
paralelas que dessincronizavam: transição não disparava para campo fora da
lista). Agora UMA declaração por propriedade na tabela `css_props!`
(style/props.rs) — campo, tipo e flags `inh`/`anim` — GERA:

- a struct ComputedStyle (docs preservados)
- merge_over (precedência da cascade; Edges mescla por lado)
- inherit_from (herança CSS — antes fn local no dom.rs)
- differs_animated (gatilho de transition — antes lista fixa com bug latente)
- interpolate_animated (lerp por TIPO via trait style::lerp::AnimValue —
  antes campo-a-campo no anim.rs)

Adicionar `opacity` agora = 1 linha na tabela + braço de parse/fmt + consumo
no paint. Mesmo espírito data-driven do Registry/SPECS do codegen.

Na mesma mexida, style.rs (2622 linhas) virou pasta style/ com submódulos
coesos (<500 linhas cada): props (tabela) / values / parse / fmt / color /
selector / stylesheet / lerp / tests. API pública IDÊNTICA via reexports do
mod.rs — zero mudança nos consumidores (layout.rs, abi.rs, rts-egui).
`is_ident` (dead code com warning) deletado.

Validação: cargo test -p rts-dom 151/151 (150 pré-existentes intactos + 1
novo cobrindo os mecanismos gerados); rts-egui compila; warnings restantes
(4) são pré-existentes de layout.rs/dom.rs, fora do escopo.
